+++ rtl/asic/soc_core.sv
@@ -219,9 +219,11 @@ module soc_core #(
         .ENABLE_IRQ        (1),
         .ENABLE_IRQ_QREGS  (1),
         .ENABLE_IRQ_TIMER  (1),
+        .ENABLE_MUL        (1),               // RV32M: multiply
+        .ENABLE_DIV        (1),               // RV32M: divide/remainder
         .PROGADDR_RESET    (32'h0000_0000),    // Boot ROM reset vector
         .PROGADDR_IRQ      (PROGADDR_IRQ),    // IRQ vector in SRAM (linker: .org 0x10 in SRAM)
-        .STACKADDR         (32'h0001_3FFC)     // Top of 16KB SRAM
+        .STACKADDR         (32'h0001_7FFC)     // Top of 32KB SRAM
     ) u_cpu (
         .clk            (clk_i),
         .resetn         (rst_ni),
@@ -374,13 +376,13 @@ module soc_core #(
     );
 
     sram_axi #(
-        .ADDR_WIDTH(14),
+        .ADDR_WIDTH(15),
         .DATA_WIDTH(32),
         .SRAM_INIT_FILE(SRAM_INIT_FILE)
     ) u_sram (
         .clk         (clk_i),
         .resetn      (rst_ni),
-        .axi_awaddr  (s1_awaddr[13:0]),
+        .axi_awaddr  (s1_awaddr[14:0]),
         .axi_awprot  (s1_awprot),
         .axi_awvalid (s1_awvalid),
         .axi_awready (s1_awready),
@@ -391,7 +393,7 @@ module soc_core #(
         .axi_bvalid  (s1_bvalid),
         .axi_bready  (s1_bready),
         .axi_bresp   (s1_bresp),
-        .axi_araddr  (s1_araddr[13:0]),
+        .axi_araddr  (s1_araddr[14:0]),
         .axi_arprot  (s1_arprot),
         .axi_arvalid (s1_arvalid),
         .axi_arready (s1_arready),